fix: Fix realizeLink func
